feat: refined release scripts
