+++ scripts/create-template-component.js
@@ -4,10 +4,16 @@
 const { Command } = require('commander');
 const fs= require('fs');
 const { mkdir, readdir, copyFile, rename } = require('fs/promises');
-const { versions } = require('process')
 const path = require('path');
-const chalk = require('chalk');
 const replaceInFile = require('replace-in-file');
+
+const {
+  warning,
+  errorLog,
+  successLog,
+  convertFirstLetterUpper,
+  isAvailableVersion,
+} = require('./helper');
 
 const program = new Command();
 const componentsDir = path.resolve(__dirname, '../src/components');
@@ -66,37 +72,3 @@ async function createTempComponent(cName) {
     errorLog(error);
   }
 }
-
-function warning(msg) {
-  if (!msg) return;
-
-  console.log(chalk.yellow(`Warning: ${msg}`));
-}
-
-function errorLog(msg) {
-  if (!msg) return;
-
-  console.log(chalk.red(`Error: ${msg}`));
-}
-
-function successLog(msg) {
-  if (!msg) return;
-
-  console.log(chalk.green(`Success🧩: ${msg}`));
-}
-
-function convertFirstLetterUpper(str = '') {
-  if (!str || typeof str !== 'string') return;
-
-  return str.charAt(0).toUpperCase() + str.slice(1);
-}
-
-function isAvailableVersion(version) {
-  if (versions.node < version) {
-    warning(`The node.js version can not less than v${version}!`)
-    return false;
-  }
-
-  return true;
-}
-
